Settings admin pages
Better settings file partitioning and consistent file naming.
