+++ admin/class-settings-fields-dev-tools.php
@@ -1,10 +1,6 @@
 <?php
 /**
- * Development admin page.
- * 
- * This class adds an admin page under Tools in the admin menu
- * at which several tools for the website development process
- * are provieded.
+ * Settings fields for site development.
  *
  * @package    Controlled_Chaos_Plugin
  * @subpackage Admin
@@ -13,7 +9,6 @@
  * @author     Greg Sweet <greg@ccdzine.com>
  * 
  * @todo       Add a "Dev Mode", functionality to be determined.
- *             Currently only contains the RTL tester.
  * @todo       Finish converting the debug plugin to work with a setting.
  */
 
@@ -25,12 +20,12 @@ if ( ! defined( 'WPINC' ) ) {
 }
 
 /**
- * Development admin page.
+ * Settings fields for site development.
  *
  * @since  1.0.0
  * @access public
  */
-class Admin_Tools {
+class Settings_Fields_Dev_Tools {
 
 	/**
 	 * Get an instance of the plugin class.
@@ -65,111 +60,9 @@ class Admin_Tools {
 	 */
 	public function __construct() {
 
-		// Add development subpage.
-		add_action( 'admin_menu', [ $this, 'dev_tools_page' ] );
-
 		// Start settings for page.
 		add_action( 'admin_init', [ $this, 'dev_settings' ] );
 		
-	}
-
-	/**
-	 * Add development subpage to Tools in the admin menu.
-	 * 
-	 * Uses the universal slug partial for admin pages. Set this
-     * slug in the core plugin file.
-	 * 
-	 * Adds a contextual help section.
-	 * 
-	 * @since  1.0.0
-	 * @access public
-	 * @return void
-	 */
-	public function dev_tools_page() {
-
-		$this->help_development = add_submenu_page( 
-			'tools.php', 
-			__( 'Website Development', 'controlled-chaos-plugin' ), 
-			__( 'Site Development', 'controlled-chaos-plugin' ), 
-			'manage_options', 
-			CCP_ADMIN_SLUG . '-dev-tools', 
-			[ $this, 'dev_tools_output' ] 
-		);
-
-		// Add content to the Help tab.
-		add_action( 'load-' . $this->help_development, [ $this, 'help_development' ] );
-
-	}
-
-	/**
-	 * Get development subpage output.
-	 * 
-	 * @since  1.0.0
-	 * @access public
-	 * @return void
-	 */
-	public function dev_tools_output() {
-
-		require plugin_dir_path( __FILE__ ) . 'partials/settings-page-development.php';
-
-	}
-
-	/**
-     * Output for the development page contextual help section.
-	 * 
-	 * @since  1.0.0
-	 * @access public
-	 * @return void
-     */
-    public function help_development() {
-
-		// Add to the development page.
-		$screen = get_current_screen();
-		if ( $screen->id != $this->help_development ) {
-			return;
-		}
-		
-		// More information.
-		$screen->add_help_tab( [
-			'id'       => 'help_dev_info',
-			'title'    => __( 'More Information', 'controlled-chaos-plugin' ),
-			'content'  => null,
-			'callback' => [ $this, 'help_dev_info_output' ]
-		] );
-		
-		// Add a help sidebar.
-		$screen->set_help_sidebar(
-			$this->help_dev_info_sidebar()
-		);
-		
-    }
-    
-    /**
-     * Get more information help tab content.
-	 * 
-	 * @since  1.0.0
-	 * @access public
-	 * @return void
-     */
-	public function help_dev_info_output() { 
-		
-		include_once plugin_dir_path( __FILE__ ) . 'partials/help/help-dev-info.php';
-	
-    }
-    
-    /**
-     * Get development page contextual tab sidebar content.
-	 * 
-	 * @since  1.0.0
-	 * @access public
-	 * @return void
-     */
-    public function help_dev_info_sidebar() {
-
-		$html = '';
-
-		return $html;
-
 	}
 
 	/**
@@ -336,11 +229,11 @@ class Admin_Tools {
  * @access public
  * @return object Returns an instance of the class.
  */
-function ccp_admin_tools() {
+function ccp_settings_fields_dev_tools() {
 
-	return Admin_Tools::instance();
+	return Settings_Fields_Dev_Tools::instance();
 
 }
 
 // Run an instance of the class.
-ccp_admin_tools();
+ccp_settings_fields_dev_tools();
